--- src/main/java/gmbs/model/InputValidator.java
@@ -1,36 +1,0 @@
-package gmbs.model;
-
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-public class InputValidator {
-
-    private static final int LENGTH_THRESHOLD = 5;
-    private static final String NATURAL_NUMBER_REGULAR_EXPRESSION = "^[0-9]*$";
-
-    private boolean isNameOverflow(String input) {
-        return input.length() > LENGTH_THRESHOLD;
-    }
-
-    private boolean hasBlank(String input) {
-        return !input.replace(" ", "").equals(input);
-    }
-
-    public boolean isValidNames(List<String> inputNames) {
-        Set<Boolean> invalidNameChecker = new HashSet<>();
-        for (String name : inputNames) {
-            invalidNameChecker.add(isNameOverflow(name));
-            invalidNameChecker.add(hasBlank(name));
-            invalidNameChecker.add(name.isEmpty());
-        }
-        return !invalidNameChecker.contains(true) && !inputNames.isEmpty();
-    }
-
-    public boolean isNaturalNumber(String userRepetitionInput) {
-        if (userRepetitionInput.matches(NATURAL_NUMBER_REGULAR_EXPRESSION)) {
-            return !userRepetitionInput.equals("0");
-        }
-        return false;
-    }
-}
